Create coords.as
Slap my wrist for putting this here. Well... where else? This is an adjunck to questspage which attempts to display active quests in the parent hold of YouAreHere. As the scaled player co-ords and scaling parms are not accessible through flash, it is not an easy task using this method.