Adds ability to "switch tabs" in the container menu
This just means the player can now switch to put items into a container.
In other versions of the game, this function can be bound to a key to be
processed in the UI. However, the VR version of the game doesn't send
a whole lot of keycodes. This makes it difficult to bind actions to
arbitrary keys on the wand/vr-controller.

This commit implements a workaround where the player can move to the
left most item on the category list: "All". When left swipe is issued
again, instead of wrapping around to the last item in the category list,
we toggle the "tab" or the direction we're operating on the container.

This at least makes the menu usable in VR.