Tweaks the position of the inventory list
Most of the information seem heavily loaded towards the left edge. This
makes using the UI quite inconvenient. Whenever the list appears, the
player needs to turn their head quite a bit to the left.

This commit moves the list towards the middle quite a bit more. This
also has the side effect of the list being closer to the item card so
looking at the itemcard does not require as much head movement.

It'd be really nice to be able to push the UI away from the player a
little bit more. An initial look at the available INIs suggest the
variable isn't exposed for tweaking.