Fixes itemcard defaulting to dragonfont *again*
It's not clear why TextFields are defaulting to use the dragon shout
font by default. Last time, some amount of messing around with the fla
file seems to eventually fix this. Apparently, whatever was done was
either reverted by accident or was not stable enough to survive
mutiple saves to itemcard.fla.

It appears things mostly break when the TextField's htmltext is updated.
For some reason, after the assignment, an extra font change tag gets
wrapped around the entire contents that turns the text into dragonfont.
It's not clear *why* this is happening. It looks like the dragonfont is
only ever used *once* in the fla.

It does look like of the multiple imported fonts in itemcard.fla,
dragonfont seems to be listed first. I wasn't able to find any
information regarding *how* to set the default.

In any case, redirecting text assignment to SetText seems to fix issue.