Updated InvalidateItem() to find multiple appearances
InvalidateItem() only invalidates the first appearances of an item in the _itemIds list. If the item appears again in the list, the later appearances are not invalidated. This leads to 'ghost' items that cause problems with other parts of the script, such as ProcessItem().

e.g. An invalid item, a_itemId, appears in both Group 1 @ index 2 and Group 4 @ index 96. 
_itemIds1.Find(a_itemId) returns 2 and sets _itemInvalidFlags1[2] = true. Processing stops and the instance in group 4 is not detected.